move game code to a join game dialog. Standin username field
