+++ src/app/app.module.ts
@@ -1,6 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
+import { ReactiveFormsModule } from '@angular/forms';
 
 import { AppComponent } from './app.component';
 import { GameRoomComponent } from './game-room/game-room.component';
@@ -10,7 +11,8 @@ import { OverlayModule } from '@angular/cdk/overlay';
 import { MatDialogModule } from '@angular/material/dialog';
 import { MatSelectModule } from '@angular/material/select';
 import { MatCheckboxModule } from '@angular/material/checkbox';
-
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { PlayingCardComponent } from './playing-card/playing-card.component';
 import { DealTaskDialogComponent } from './deal-task-dialog/deal-task-dialog.component';
@@ -19,6 +21,7 @@ import { InformationCardsComponent } from './communication-cards/information-car
 import { GameSummaryComponent } from './game-summary/game-summary.component';
 
 import { GameService } from './services/game.service';
+import { JoinGameDialogComponent } from './join-game-dialog/join-game-dialog.component';
 
 @NgModule({
   declarations: [
@@ -29,6 +32,7 @@ import { GameService } from './services/game.service';
     TaskSelectionComponent,
     InformationCardsComponent,
     GameSummaryComponent,
+    JoinGameDialogComponent,
   ],
   imports: [
     BrowserModule,
@@ -40,6 +44,9 @@ import { GameService } from './services/game.service';
     DragDropModule,
     MatSelectModule,
     MatCheckboxModule,
+    ReactiveFormsModule,
+    MatFormFieldModule,
+    MatInputModule,
   ],
   providers: [GameService],
   bootstrap: [AppComponent],
